refactor(cli): Moves remaining validation/services/ Moves remaining code in `services/validation/` to `command/check/utils/`
