+++ cli/src/commands/check/utils/check-schema-compatibility.ts
