+++ cli/src/commands/check/utils/flatten-schemas.ts
